pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot"
version = "0.1.0"
description = "RAG-based Research Assistant Chatbot"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gradio>=6.2.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.0",
    "pinecone-client>=6.0.0",
    "pymongo>=4.15.5",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["chatbot*"]
